Create Room Form moved to be a component
